# data-api/WoW/schemas/wow_domain_swagger_2.0.yaml
swagger: "2.0"
info:
    title: weather_on_the_web_core
    description: 'Core methods for Weather on the Web API''s'
    version: 1.0.0
host: data.example.org
basePath: /wotw
schemes:
  - https
  - http
paths:
...
    /groups:
        get:
            tags:
                - Capabilities
            summary: 'Provide a list of avialable data groups'
            operationId: listGroups
            responses:
                200:
                    description: 'links to the API capabilities'
                    schema:
                      $ref: '#/definitions/content'
                default:
                    description: 'An error occured.'
                    schema:
                      $ref: '#/definitions/exception'
    '/groups/{groupId}':
        get:
            tags:
                - Capabilities
            summary: 'List of links to information available in the group'
            operationId: groupInfomation
            responses:
                200:
                    description: 'links to the API groups'
                    schema:
                      $ref: '#/definitions/content'
                default:
                    description: 'An error occured.'
                    schema:
                      $ref: '#/definitions/exception'    
        parameters:
            -
                name: groupId
                in: path
                required: true
                description: 'Identifier (name) of a specific group'
                type: string
            -
                in: query
                name: bbox
                description: "Only features that have a geometry that intersects the bounding box are selected. The bounding box is provided as four or six numbers, depending on whether the coordinate reference system includes a vertical axis (elevation or depth): * Lower left corner, coordinate axis 1 * Lower left corner, coordinate axis 2 * Lower left corner, coordinate axis 3 (optional) * Upper right corner, coordinate axis 1 * Upper right corner, coordinate axis 2 * Upper right corner, coordinate axis 3 (optional) The coordinate reference system of the values is WGS84 longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless a different coordinate reference system is specified in the parameter `bbox-crs`. For WGS84 longitude/latitude the values are in most cases the sequence of minimum longitude, minimum latitude, maximum longitude and maximum latitude. However, in cases where the box spans the antimeridian the first value (west-most box edge) is larger than the third value (east-most box edge). If a feature has multiple spatial geometry properties, it is the decision of the server whether only a single spatial geometry property is used to determine the extent or all relevant geometries. "
                maxItems: 6
                minItems: 4
                type: array
                items:
                    type: number
            -
                name: time
                in: query
                description: "Either a date-time or a period string that adheres to RFC 3339. Examples: * A date-time:   2018-02-12T23:20:50Z   * A period:   2018-02-12T00:00:00Z/2018-03-18T12:31:12Z   or   2018-02-12T00:00:00Z/P1M6DT12H31M12S   Only features that have a temporal property that intersects the value of `time` are selected. If a feature has multiple temporal properties, it is the decision of the server whether only a single temporal property is used to determine the extent or all relevant temporal properties."
                type: string
    /collections:
        get:
            tags:
                - Capabilities
            summary: 'List the avialable collections from the service'
            operationId: listCollections
            responses:
                200:
                    description: 'links to a list of collections'
                    schema:
                      $ref: '#/definitions/content'
                default:
                    description: 'An error occured.'
                    schema:
                      $ref: '#/definitions/exception'
        parameters:
            -
                in: query
                name: bbox
                description: "Only features that have a geometry that intersects the bounding box are selected. The bounding box is provided as four or six numbers, depending on whether the coordinate reference system includes a vertical axis (elevation or depth): * Lower left corner, coordinate axis 1 * Lower left corner, coordinate axis 2 * Lower left corner, coordinate axis 3 (optional) * Upper right corner, coordinate axis 1 * Upper right corner, coordinate axis 2 * Upper right corner, coordinate axis 3 (optional) The coordinate reference system of the values is WGS84 longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless a different coordinate reference system is specified in the parameter `bbox-crs`. For WGS84 longitude/latitude the values are in most cases the sequence of minimum longitude, minimum latitude, maximum longitude and maximum latitude. However, in cases where the box spans the antimeridian the first value (west-most box edge) is larger than the third value (east-most box edge). If a feature has multiple spatial geometry properties, it is the decision of the server whether only a single spatial geometry property is used to determine the extent or all relevant geometries. "
                maxItems: 6
                minItems: 4
                type: array
                items:
                    type: number
            -
                name: time
                in: query
                description: "Either a date-time or a period string that adheres to RFC 3339. Examples: * A date-time:   2018-02-12T23:20:50Z   * A period:   2018-02-12T00:00:00Z/2018-03-18T12:31:12Z   or   2018-02-12T00:00:00Z/P1M6DT12H31M12S   Only features that have a temporal property that intersects the value of `time` are selected. If a feature has multiple temporal properties, it is the decision of the server whether only a single temporal property is used to determine the extent or all relevant temporal properties."
                type: string
    '/collections/{collectionId}':
        get:
            tags:
                - Metadata
            summary: 'list data instances of  {collectionId}'
            description: 'This will provide information about the available instances for the chosen collection Use content negotiation to request HTML or GeoJSON.'
            operationId: getCollection
            responses:
                200:
                    description: 'Metadata about the {collectionId} collection shared by this API.'
                    schema:
                      $ref: '#/definitions/content'
                default:
                    description: 'An error occured.'
                    schema:
                      $ref: '#/definitions/exception'
        parameters:
            -
              name: collectionId
              in: path
              description: 'Identifier (name) of a specific collection'
              required: true
              type: string
    '/collections/{collectionId}/{instanceId}':
        get:
            tags:
                - Metadata
            summary: 'list data instances of  {collectionId}'
            description: 'This will provide metadata for the chosen instance of a collection Use content negotiation to request HTML or GeoJSON.'
            operationId: getCollectionInstance
            responses:
                200:
                    description: 'Metadata about the instance of  {collectionId} collection shared by this API.'
                    schema:
                        $ref: '#/definitions/collectionInfo'
                default:
                    description: 'An error occured.'
                    schema:
                      $ref: '#/definitions/exception'
        parameters:
            -
              name: collectionId
              in: path
              description: 'Identifier (name) of a specific collection'
              required: true
              type: string
            -
              name: instanceId
              in: path
              description: 'Identifier (name) of a specific instance of a collection'
              required: true
              type: string
    '/collections/{collectionId}/{instanceId}/point':
        get:
            tags:
                - Data
            summary: 'Query end point for point queries of instance {instanceId} of collection {collectionId}'
            description: 'Query end point for point queries'
            operationId: getDataFromPoint
            parameters:
                -
                  name: collectionId
                  in: path
                  description: 'Identifier (name) of a specific collection'
                  required: true
                  type: string
                -
                  name: instanceId
                  in: path
                  description: 'Identifier (name) of a specific instance of a collection'
                  required: true
                  type: string
                -
                    name: parametername
                    in: query
                    description: comma delimited list of parameters to retrieve data for.  Valid parameters are listed in the collections metadata
                    required: true
                    type: string
                -
                    name: time
                    in: query
                    description: Period to return data for in the RFC 3339 profile of ISO 8601, datetime values MUST be defined to seconds i.e. 2019-02-05T21:15:00 (YYYY-MM-DDTHH:MM:SSZ).  A time range MUST be defined as Start_Time/End_Time i.e. (2019-02-05T21:15:00Z/2019-02-06T12:45:00Z) and time range with fixed intervals MUST be defined as Intervals/Start_Time/End_Time i.e. (R6/2019-02-05T21:15:00Z/2019-02-05T22:45:00Z would be 15 minute timesteps between 2019-02-05T21:15:00Z and 019-02-05T22:45:00Z)
                    required: true
                    format: date-time
                    type: string
                -
                    name: crs
                    in: query
                    description: 'The EPSG code for the coordinate system of the X and Y values (i.e. EPSG:3857), if not defined X and Y will default to EPSG:4326 longitude and latitude values'
                    type: string
                -
                    name: outputFormat
                    in: query
                    description: 'Data format to return the query results in (if not specified this will default to CoverageJSON)'
                    required: false
                    type: string
                -
                    name: interpolation
                    in: query
                    description: "This defines the algorithm that identifies the best data match for the requested location The valid interpolation method values are available in the results of the  capabilities query methods of this API "
                    type: string
                -
                    name: coords
                    in: query
                    description: "location(s) to return data for, the coordinates are defined by a Well Known Text (wkt) string. to retrieve a single location :  POINT(x y) i.e. POINT(0,51.48) for Greenwich, London  And for a list of locations  MULTIPOINT((x y),(x1 y1),(x2 y2),(x3 y3))  i.e. MULTIPOINT((38.9 -77),(48.85 2.35),(39.92 116.38),(-35.29 149.1),(51.5 -0.1))   see http://portal.opengeospatial.org/files/?artifact_id=25355 and https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry  the coordinate values will depend on the CRS parameter, if this is not defined the values will be assumed to WGS85 values (i.e x=longitude and y=latitude) "
                    required: true
                    type: string
            responses:
                200:
                    description: 'Data ranges required to construct valid queries for the choosen data collection'
                    schema:
                      $ref: '#/definitions/CoverageJSON'
                default:
                    description: 'An error occured.'
                    schema:
                      $ref: '#/definitions/exception'
    '/collections/{collectionId}/{instanceId}/polygon':
        get:
            tags:
                - Data
            summary: 'Query end point for area queries of instance {instanceId} of collection {collectionId} defined by a polygon'
            description: 'Return the data values for the data Polygon defined by the query parameters'
            operationId: GetDataFromArea
            parameters:
                -
                  name: collectionId
                  in: path
                  description: 'Identifier (name) of a specific collection'
                  required: true
                  type: string
                -
                  name: instanceId
                  in: path
                  description: 'Identifier (name) of a specific instance of a collection'
                  required: true
                  type: string
                -
                    name: parametername
                    in: query
                    description: "comma delimited list of parameters to retrieve data for.  Valid parameters are listed in the collections metadata"
                    required: true
                    minItems: 1
                    type: array
                    items:
                        type: string
                -
                    name: outputFormat
                    in: query
                    description: 'Data format of the returned the data values'
                    required: true
                    type: string
                -
                    name: crs
                    in: query
                    description: "Coordinate system that the data should be returned in, defined as an EPSG code  value e.g. EPSG:4326.  If not specified output will be in the native coordinate  system of the data"
                    required: false
                    type: string
                -
                    name: coords
                    in: query
                    description: "Only data that has a geometry that intersects the area defined by the polygon are selected.   The polygon is defined using a Well Known Text string following   coords=POLYGON((x y,x1 y1,x2 y2,...,xn yn x y))  which are values in the coordinate system defined by the crs query parameter  (if crs is not defined the values will be assumed to be WGS84 longitude/latitude coordinates).    For instance a polygon that roughly describes an area that contains  South West England in WGS84 would look like:   coords=POLYGON((-6.1 50.3,-4.35 51.4,-2.6 51.6,-2.8 50.6,-5.3 49.9,-6.1,50.3))  see http://portal.opengeospatial.org/files/?artifact_id=25355 and https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry  The coords parameter will only support 2D POLYGON definitions"
                    required: true
                    type: string
                -
                    name: time
                    in: query
                    description: "Either a date-time or a period string that adheres to RFC 3339, Only data that  has a temporal property that intersects the value of time are selected.    datetime values MUST be defined to seconds i.e. 2019-02-05T21:15:00 (YYYY-MM-DDTHH:MM:SSZ). A time range MUST be defined as Start_Time/End_Time i.e. (2019-02-05T21:15:00Z/2019-02-06T12:45:00Z) and time range with fixed intervals MUST be defined as Repition/Start_Time/End_Time i.e. (R3/2019-02-05T21:15:00Z/2019-02-05T22:00:00Z would be 15 minute timesteps between 2019-02-05T21:15:00Z and 019-02-05T22:00:00Z)   "
                    required: true
                    type: string
                -
                    name: interpolationX
                    in: query
                    description: "This is defined if the query doesn't require data at its native X coordinate resolution, the format is:  interpolationX=R<interpolation resolution>/<interpolation method> e.g. interpolationX=R15/linear  The interpolation resolution is the number of intervals the user requires between the minimum X value of the bounding box and the maximum X value of the bounding box.  The valid interpolation method values are available in the results of the capabilities query methods of this API  "
                    type: string
                -
                    name: interpolationY
                    in: query
                    description: "This is defined if the query doesn't require data at its native Y coordinate  resolution, the format is:  interpolationY=R<interpolation resolution>/<interpolation method>  e.g. interpolationY=R5/linear  The interpolation resolution is the number of intervals the user requires between the minimum Y value of the bounding box and the maximum Y value of the bounding box.  The valid interpolation method values are available in the results of the  capabilities query methods of this API "
                    type: string
                -
                    name: z
                    in: query
                    description: "Define the vertical level to return data from  i.e. Z=level  for instance if the 850hPa pressure level is being queried  Z=850  When not specified data from the lowest (i.e. nearest the surface) level of data in the collection. "
                    type: string
            responses:
                200:
                    description: 'A coverage'
                    schema:
                        $ref: '#/definitions/CoverageJSON'
                default:
                    description: 'An error occured.'
                    schema:
                        $ref: '#/definitions/exception'
...
definitions:
    exception:
        required:
            - code
        type: object
        properties:
            code:
                type: string
            description:
                type: string
...
    link:
        required:
            - href
        type: object
        properties:
            href:
                type: string
            rel:
                type: string
                example: prev
            type:
                type: string
                example: application/geo+json
            hreflang:
                type: string
                example: en
    content:
        required:
            - links
            - collections
        type: object
        properties:
            links:
                type: array
                items:
                    $ref: '#/definitions/link'
                example:
                    -
                        href: 'http://data.example.org/collections.json'
                        rel: self
                        type: application/json
                        title: 'this document'
                    -
                        href: 'http://data.example.org/collections.html'
                        rel: alternate
                        type: text/html
                        title: 'this document as HTML'
                    -
                        href: 'http:/.example.org/1.0/foobar.xsd'
                        rel: describedBy
                        type: application/xml
                        title: 'XML schema for Acme Corporation data'
            collections:
                type: array
                items:
                    $ref: '#/definitions/collectionInfo'
    collectionInfo:
        required:
            - name
            - links
        type: object
        properties:
            name:
                description: 'identifier of the collection used, for example, in URIs'
                type: string
                example: observations
            title:
                description: 'human readable title of the collection'
                type: string
                example: 'Observation data'
            description:
                description: 'a description of the features in the collection'
                type: string
                example: 'Observations available from the api'
            links:
                type: array
                items:
                    $ref: '#/definitions/link'
                example:
                    -
                        href: 'http://data.example.org/collections/observations'
                        rel: item
                        type: application/geo+json
                        title: 'Observation data'
                    -
                        href: 'http://example.org/concepts/building.html'
                        rel: describedBy
                        type: text/html
                        title: 'List of available observation types'
            extent:
                $ref: '#/definitions/extent'
            crs:
                description: 'The coordinate reference systems in which geometries may be retrieved. Coordinate reference systems are identified by a URI. The first coordinate reference system is the coordinate reference system that is used by default. This is always "http://www.opengis.net/def/crs/OGC/1.3/CRS84", i.e. WGS84 longitude/latitude.'
                default:
                    - 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'
                type: array
                items:
                    type: string
    extent:
        description: 'Definition of the geo-temporal extent of the collection'
        required:
            - dimensions
            - wkt
            - horizontal
        type: object
        properties:
            dimensions:
                $ref: '#/definitions/dimension'
            wkt:
                description: "Well Known Text description of the collections coordinate reference system  see http://docs.opengeospatial.org/is/12-063r5/12-063r5.html"
                type: string
            horizontal:
                $ref: '#/definitions/horizontal'
            vertical:
                $ref: '#/definitions/vertical'
            temporal:
                $ref: '#/definitions/temporal'
        example: { "dimensions": {"x": 92, "y": 33,"z": 8,"time": 98},"wkt": "GEOGCS[  WGS 84  ,DATUM[  WGS_1984  ,SPHEROID[  WGS 84  ,6378137,298.257223563,AUTHORITY[  EPSG  ,  7030  ]],AUTHORITY[  EPSG  ,  6326  ]],PRIMEM[  Greenwich  ,0,AUTHORITY[  EPSG  ,  8901  ]],UNIT[  degree  ,0.01745329251994328,AUTHORITY[  EPSG  ,  9122  ]],AUTHORITY[  EPSG  ,  4326  ],TIMEEXTENT[2019-05-28T06:00:00Z,2019-05-30T06:00:00Z],VERTICALEXTENT[2.0,2.0,LENGTHUNIT[  metre  ,1.0]]],  horizontal  : {  name  : [  longitude  ,  latitude  ],  coordinates  : [   x  ,   y  ],  geographic  :   BBOX[359.875,90.125,-0.125,-90.125]  },  vertical  : {  name  : [  height_above_ground  ],  coordinates  : [  z  ],  range  : [ 36.47, 99.07]},   temporal  : {  name  :   time  ,   coordinates  : [  time  ],    range  : [   2019-05-28T06:00:00Z  ,   2019-05-28T18:00:00Z  ,   2019-05-29T06:00:00Z  ,  2019-05-29T18:00:00Z  ,  2019-05-30T06:00:00Z  ]"}
...
    Parameter:
        title: 'Root Type for Parameter'
        description: 'Definition of data parameter'
        required:
            - type
            - unit
            - description
            - observedProperty
            - id
        type: object
        properties:
            type:
                description: type
                enum:
                    - Parameter
            description:
                $ref: '#/definitions/descriptionInfo'
            unit:
                $ref: '#/definitions/units'
            observedProperty:
                $ref: '#/definitions/ObservedProperty'
            id:
                description: 'Unique ID of the parameter, this is the value used for querying the data'
                type: string
        example: {"type" : "Parameter","id" : "sea_ice","description": {"en": "Sea Ice concentration (ice=1;no ice=0)"},"unit" : {"label": {"en": "Ratio"},"symbol": {"value": "1","type": "http://www.opengis.net/def/uom/UCUM/"}},"observedProperty" : {"id" : "http://vocab.nerc.ac.uk/standard_name/sea_ice_area_fraction/","label" : {"en": "Sea Ice Concentration"}}}
...
    CoverageJSON:
        title: 'Root Type for CoverageJSON'
        description: 'JSON output format '
        required:
            - type
            - domain
            - parameters
            - ranges
        type: object
        properties:
            type:
                description: 'Coverage domain type'
                enum:
                    - Coverage
            domain:
                $ref: '#/definitions/Domain'
                description: 'Geotemporal domain definition'
                properties:
                    type:
                        type: string
                    domainType:
                        type: string
                    axes:
                        type: object
                        properties:
                            x:
                                type: object
                                properties:
                                    values:
                                        type: array
                                        items: {}
                            y:
                                type: object
                                properties:
                                    values:
                                        type: array
                                        items: {}
                            z:
                                type: object
                                properties:
                                    values:
                                        type: array
                                        items: {}
                            t:
                                type: object
                                properties:
                                    values:
                                        type: array
                                        items: {}
                    referencing:
                        type: array
                        items: {}
            parameters:
                $ref: '#/definitions/Parameter'
                description: "A set of objects (one for each parameter) which describe the data values  contained within the JSON"
                properties:
                    ICEC:
                        type: object
                        properties:
                            type:
                                type: string
                            description:
                                type: object
                                properties:
                                    en:
                                        type: string
                            unit:
                                type: object
                                properties:
                                    label:
                                        type: object
                                        properties:
                                            en:
                                                type: string
                                    symbol:
                                        type: object
                                        properties:
                                            value:
                                                type: string
                                            type:
                                                type: string
                            observedProperty:
                                type: object
                                properties:
                                    id:
                                        type: string
                                    label:
                                        type: object
                                        properties:
                                            en:
                                                type: string
            ranges:
                $ref: '#/definitions/range'
                description: 'A set of objects (one for each parameter) which contain the data values'
                properties:
                    ICEC:
                        type: object
                        properties:
                            type:
                                type: string
                            dataType:
                                type: string
                            axisNames:
                                type: array
                                items: {}
                            shape:
                                type: array
                                items: {}
                            values:
                                type: array
                                items: {}
...